Separate chat header into different component
